add cart icon to each DeviceItem.jsx and ability to add this device to cart directly from shop. Create auto countable rating(needs FIX to show decimal nums) after new review. Sort by price. Clean code in deviceController.js getAll()

+++ server/model/Device.js
@@ -5,7 +5,7 @@ module.exports = sequelize.define('device', {
     id: {type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true},
     name: {type: DataTypes.STRING, unique: true, allowNull: false},
     price: {type: DataTypes.INTEGER, allowNull: false},
-    rating: {type: DataTypes.INTEGER, defaultValue: 0},
+    rating: {type: DataTypes.DOUBLE, defaultValue: 0},
     img: {type: DataTypes.STRING, allowNull: false},
     brandId: { type: DataTypes.INTEGER, allowNull: false },
     brandName: { type: DataTypes.STRING, allowNull: false },
